Move AppDomainConfiguration.cs to Plumbing
