+++ source/Calamari.Common/Plumbing/AppDomainConfiguration.cs
@@ -1,13 +1,13 @@
 using System;
 
-namespace Calamari.Common
+namespace Calamari.Common.Plumbing
 {
     internal static class AppDomainConfiguration
     {
         public static readonly TimeSpan DefaultRegexMatchTimeout = TimeSpan.FromSeconds(1);
         
         /// <summary>
-        /// Adds default max timout to avoid ReDoS attacks
+        /// Adds default max timeout to avoid ReDoS attacks
         /// </summary>
         public static void SetDefaultRegexMatchTimeout() 
             => AppDomain.CurrentDomain.SetData("REGEX_DEFAULT_MATCH_TIMEOUT", DefaultRegexMatchTimeout);
